refactor(anti-fraud): layered architecture implemented
